Host of Swift 4 changes and updates. Also Source -> Sources in root dir
