+++ Sources/Starscream.swift
@@ -3,7 +3,7 @@
 //
 // Based on commit SHA 789264eeff101e, but without some docs fixes from ee993322c
 // onwards, and without the compression support.
-
+//
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
 //  SSLSecurity.swift
